Uncomment open() and close() methods for BaseOpenCloseElement
Because expand() and collapse() are used to wrap open() and close(), it's best to keep them all defined, even though not implemented. If a subclass defines open() and close(), it would also need to define expand() and collapse() if they were commented out in BaseOpenCloseElement. Uncommenting them gets you the wrapper functions 'for free.'